Move ActivateResult to application package
