--- controller-server/src/main/java/com/yahoo/vespa/hosted/controller/api/package-info.java
@@ -1,8 +1,0 @@
-// Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-/**
- * @author Tony Vaagenes
- */
-@ExportPackage
-package com.yahoo.vespa.hosted.controller.api;
-
-import com.yahoo.osgi.annotation.ExportPackage;
