+++ controller-server/src/main/java/com/yahoo/vespa/hosted/controller/application/ActivateResult.java
@@ -1,5 +1,5 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.hosted.controller.api;
+package com.yahoo.vespa.hosted.controller.application;
 
 import com.yahoo.vespa.hosted.controller.api.identifiers.RevisionId;
 import com.yahoo.vespa.hosted.controller.api.integration.configserver.PrepareResponse;
